jme3-alloc: added final java-docs
